third_party/blink/web_tests/platform/mac-mac10.15/virtual/plz-service-worker/external/wpt/service-workers/service-worker/xhr-content-length.https.window-expected.txt

This is a testharness.js-based test.
PASS Setup
PASS Synthetic response without Content-Length header
PASS Synthetic response with Content-Length header with value larger than response body length
FAIL Synthetic response with two Content-Length headers value larger than response body length assert_equals: expected "10000, 10000" but got "10000"
PASS Synthetic response with bogus Content-Length header
Harness: the test ran to completion.
